+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `configs` section to `App` CRD based on RFC: https://github.com/giantswarm/rfc/tree/main/multi-layer-app-config#enhancing-app-cr
+
 ## [0.4.1] - 2022-06-30
 
 ## [0.4.1] - 2022-06-30